+++ !SourceCode/SRC LZSS.exe/V1_02/src/JB01_Decompress.h
@@ -117,8 +117,8 @@ public:
 	ULONG	GetDecompressedSize(void);
 
 	// Monitor functions
-	UINT	GetPercentComplete(void) { return ((UINT)(((double)m_nUserDataPos/(double)m_nDataSize) * 100.0)); }
-
+//	UINT	GetPercentComplete(void) { return ((UINT)(((double)m_nUserDataPos/(double)m_nDataSize) * 100.0)); }
+	UINT	GetPercentComplete(void) { return ((m_nUserDataPos / m_nDataSize) * 100); }
 private:
 	// User supplied buffers and counters
 	UCHAR	*m_bUserData;						// When compressing from memory this is the user input buffer
